Equation Classes
Set up a class hierarchy system. Value is the base class and classes such as Operator, Constant, Variable, and Expression extend upon this class.  The Function class extends upon Expression. Upon calling the constructor of the Expression class a String is taken and analyzed splitting it apart and parsing it into a list of extensions of the Value class. This constructor of the Expression class tends to be recursive as there are often nested Expressions and Functions within a single Expression or Function. Once an Expression has completed it's construction it is now encapsulated and organized in a form to be solved using the solve() method. Any unassigned variables denoted by alphabetic letters are automatically assigned as zero, though there is a quick way to dynamically assign values to alphabetic letters through the Variable.variableValue enumerator. Once the solve function is called within the Expression or Function class the expression is solved using mathematical convention. Any invalid formation of an Expression such as through repeated mathematical operators or empty parentheses throw exceptions.