Calculate Click
In game modification of the graphed equation is complete. The user can now specify the equation that is graphed while using the program and specify the scale and bounds of the graph. There is still some work needed to be done on possible debugging, input control, instructional information, and ease of use such as glitches in the textboxes.